MYR-805: renamed friend to friendsMenu
