+++ src/components-v2/FriendsMenu/friend-list.tsx
